Ajustes no carrinho

+++ src/main/kotlin/br/com/unip/carrinho/dto/CarrinhoDTO.kt
@@ -1,11 +1,10 @@
 package br.com.unip.carrinho.dto
 
-import java.math.BigDecimal
 import java.time.LocalDateTime
 
 class CarrinhoDTO(val id: String,
                   val produtos: List<ProdutoCarrinhoDTO>,
-                  val valorTotal: BigDecimal,
+                  val valorTotal: Double,
                   val dataCriacao: LocalDateTime,
                   val fornecedorUUID: String?)
