+++ src/main/kotlin/br/com/unip/carrinho/dto/ItemDTO.kt
@@ -1,5 +1,3 @@
 package br.com.unip.carrinho.dto
 
-import java.math.BigDecimal
-
-class ItemDTO(val id: String, val nome: String, val observacoes: String?, val quantidade: Long, val valor: BigDecimal)
+class ItemDTO(val id: String, val nome: String, val observacoes: String?, val quantidade: Long, val valor: Double)
