+++ src/main/kotlin/br/com/unip/carrinho/dto/PedidoDTO.kt
@@ -1,12 +1,11 @@
 package br.com.unip.carrinho.dto
 
-import java.math.BigDecimal
 import java.time.LocalDateTime
 
 class PedidoDTO(val id: String,
                 val numero: String,
                 var itens: List<ItemDTO> = emptyList(),
                 val status: String,
-                val valor: BigDecimal,
+                val valor: Double,
                 val cliente: ClienteDTO,
                 val dataPedido: LocalDateTime)
