+++ src/main/kotlin/br/com/unip/carrinho/dto/ProdutoDTO.kt
@@ -1,14 +1,12 @@
 package br.com.unip.carrinho.dto
 
-import java.math.BigDecimal
-
 class ProdutoDTO {
 
     var id: String
     var nome: String
-    var valor: BigDecimal
+    var valor: Double
 
-    constructor(id: String, nome: String, valor: BigDecimal) {
+    constructor(id: String, nome: String, valor: Double) {
         this.id = id
         this.nome = nome
         this.valor = valor
