+++ src/main/kotlin/br/com/unip/carrinho/repository/entity/Pedido.kt
@@ -1,12 +1,9 @@
 package br.com.unip.carrinho.repository.entity
 
 import br.com.unip.carrinho.repository.entity.enums.EStatusPedido
-import br.com.unip.carrinho.repository.entity.enums.EStatusPedido.CONCLUIDO
-import br.com.unip.carrinho.repository.entity.enums.EStatusPedido.PENDENTE_PAGAMENTO
-import br.com.unip.carrinho.repository.entity.enums.EStatusPedido.PENDENTE_PREPARACAO
+import br.com.unip.carrinho.repository.entity.enums.EStatusPedido.*
 import org.springframework.data.annotation.Id
 import org.springframework.data.mongodb.core.mapping.Document
-import java.math.BigDecimal
 import java.time.LocalDateTime
 
 @Document(collection = "pedido")
@@ -25,7 +22,7 @@ class Pedido {
 
     var status: EStatusPedido = PENDENTE_PAGAMENTO
 
-    var valor: BigDecimal
+    var valor: Double
 
     lateinit var pagamento: Pagamento
 
@@ -33,7 +30,7 @@ class Pedido {
 
     var cliente: Cliente
 
-    constructor(fornecedorUUID: String, cadastroUUID: String, numero: String, itens: List<Item>, valor: BigDecimal, cliente: Cliente) {
+    constructor(fornecedorUUID: String, cadastroUUID: String, numero: String, itens: List<Item>, valor: Double, cliente: Cliente) {
         this.fornecedorUUID = fornecedorUUID
         this.cadastroUUID = cadastroUUID
         this.numero = numero
