+++ src/main/kotlin/br/com/unip/carrinho/webservice/model/response/CarrinhoResponse.kt
@@ -3,7 +3,6 @@ package br.com.unip.carrinho.webservice.model.response
 import com.fasterxml.jackson.annotation.JsonFormat
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonProperty
-import java.math.BigDecimal
 import java.time.LocalDateTime
 
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -16,13 +15,13 @@ class CarrinhoResponse {
     val itens: List<ItemCarrinhoResponse>
 
     @JsonProperty(value = "valor_total")
-    val valorTotal: BigDecimal
+    val valorTotal: Double
 
     @JsonProperty(value = "data_criacao")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy HH:mm:ss")
     var dataCriacao: LocalDateTime
 
-    constructor(id: String?, itens: List<ItemCarrinhoResponse>, valorTotal: BigDecimal, dataCriacao: LocalDateTime) {
+    constructor(id: String?, itens: List<ItemCarrinhoResponse>, valorTotal: Double, dataCriacao: LocalDateTime) {
         this.id = id
         this.itens = itens
         this.valorTotal = valorTotal
